Agregar sensores termometro y suelo al BLE Agregar letrero de sensor no coenctado en sensor suelo Arreglar deep sleep cuando está en patanlla de lenguaje

+++ include/hw.h
@@ -40,9 +40,10 @@ void init_hw();
 int read_sound_level();
 
 /**
- * @brief Lee la humedad del suelo y devuelve porcentaje (0-100) calibrado
+ * @brief Lee la humedad del suelo y devuelve porcentaje (0-100) calibrado.
+ * Devuelve NAN si el pin parece flotante y el sensor probablemente no está conectado.
  */
-int read_soil_moisture();
+float read_soil_moisture();
 
 /**
  * @brief Lee temperatura en °C del sensor DS18B20
